Fix graph metadata lookups
NetworkX allows nodes to have metadata, like so:

    graph.node[id][key] = value

The CiteNet codebase was incorrectly using:

    graph[id][key] = value

This second syntax actually creates a new edge from `id` to `key`, which
is totally unexpected and incorrect behavior. These phantom edges were
changing the results!

This problem was discovered when the report outputs changed after new
metadata was added. With the fixes in this commit, the report data no
longer changes with different metadata inputs.